simplified submit, complete, finish buttons
