+++ src/components/form/FTextField.tsx
@@ -1,17 +1,19 @@
-import { useFormContext, Controller } from "react-hook-form";
+import { useFormContext, Controller, RegisterOptions } from "react-hook-form";
 import { TextField, TextFieldProps } from "@mui/material";
 type FTextFieldProps = {
   name: string;
   defaultValue?: any;
+  rule?: RegisterOptions;
 } & TextFieldProps;
 
-function FTextField({ name, defaultValue, ...other }: FTextFieldProps) {
-  const { control } = useFormContext();
+function FTextField({ name, defaultValue, rule, ...other }: FTextFieldProps) {
+  const { control, formState } = useFormContext();
   return (
     <Controller
       name={name}
       control={control}
-      defaultValue={defaultValue || ""}
+      defaultValue={defaultValue}
+      rules={rule}
       render={({ field, fieldState: { error } }) => (
         <TextField
           {...field}
